<refactor> : step4 자동차 경주(우승자) 리뷰 반영 리팩토링(3)
