+++ src/main/java/racingcargame/Distance.java
@@ -7,19 +7,11 @@ public class Distance {
         this.value = value;
     }
 
-    public Distance(RacingCar racingCar) {
-        value = racingCar.showDistance();
-    }
-
     public int increase() {
         return value += 1;
     }
 
     public int getValue() {
         return value;
-    }
-
-    public int showWinnerDistance(RacingCar racingCar) {
-        return racingCar.showDistance();
     }
 }
