+++ src/main/java/racingcargame/RacingCarName.java
@@ -3,16 +3,8 @@ package racingcargame;
 public class RacingCarName {
     private String name;
 
-    RacingCarName(RacingCar racingCar) {
-        this.name = racingCar.showCarName();
-    }
-
     RacingCarName(String name) {
         this.name = name;
-    }
-
-    public String showWinnerName(RacingCar racingCar) {
-        return name;
     }
 
     public String getName() {
